### 2024-05-23 #6770 @editorialbot
[PRE REVIEW]: UNSAFE: An UNcertain Structure And Fragility Ensemble framework for property-level flood risk estimation
**Submitting author:** @abpoll (<a href="http://orcid.org/0000-0001-6642-0591">Adam Pollack</a>)
**Repository:** https://github.com/abpoll/unsafe
**Branch with paper.md** (empty if default branch):
**Version:** v0.1
