### 2024-06-06 #6770 @atrisovic
Hi @erexer,
I am writing to invite you to review a paper titled "UNSAFE: An UNcertain Structure And Fragility Ensemble framework for property-level flood risk estimation", which is currently in the review process at the Journal of Open Source Software (JOSS). Your insights and feedback on this work would be valuable and much appreciated. To help with the review, we have [a reviewers' checklist](https://joss.readthedocs.io/en/latest/review_checklist.html#review-checklist) with useful guidelines.
Thank you!
Ana Trisovic, JOSS Topic Editor
