### 2024-11-04 #6770 @atrisovic
Hi All,

@mebauer @javedali99, @pkjr002 @mshiv,

I wanted to follow up on my previous invitation to review the paper titled "UNSAFE: An UNcertain Structure And Fragility Ensemble framework for property-level flood risk estimation" currently under review at the Journal of Open Source Software (JOSS). Your insights and expertise would be incredibly valuable to the evaluation of this work, and we would greatly appreciate your feedback. If you're available and interested, we would be glad to share the reviewers' checklist and any additional guidelines to assist with the process.

Please let us know if you might be available to take on this review.
