+++ joss_submission/paper.md
@@ -50,7 +50,7 @@ Figure 1 illustrates how accounting for uncertainty in exposure and vulnerabilit
 ![**Demonstration of the UNSAFE approach.** Accounting for uncertainty in exposure and vulnerability can result in a large range of plausible flood damage estimates and increases the expected value of risk. The red vertical line shows the expected annual damage estimate produced from the standard HAZUS approach, which neglects uncertainty in depth-damage functions and national structure inventory records. The orange histogram shows the range of plausible damage estimates when accounting for uncertainty in HAZUS depth-damage functions and national structure inventory records. The blue histogram shows the same, but with uncertainty in NACCS depth-damage functions instead of HAZUS ones. At the top, orange and blue boxplots summarize the distributions of HAZUS and NACCS losses, respectively. The red diamond denotes the ensemble mean for each distribution. \label{fig:philly_results}](philly_results.png)
 
 # Acknowledgements
-This work was co-supported by the Earth System Model Development and Regional and Global Modeling and Analysis program areas of the U.S. Department of Energy, Office of Science, Office of Biological and Environmental Research as part of the multi-program, collaborative Integrated Coastal Modeling (ICoM) project and the Thayer School of Engineering. All errors and opinions are those of the authors and not of the supporting entities.
+This research was supported by the U.S. Department of Energy, Office of Science, as part of research in MultiSector Dynamics, Earth and Environmental System Modeling Program and Dartmouth College. All errors and opinions are those of the authors and not of the supporting entities.
 
 Software License: `UNSAFE` is distributed under the BSD-2-Clause license. The authors do not assume responsibility for any (mis)use of the provided code.
 
