### 2024-12-04 421e5fc
fix typo in environment.yml
