+++ examples/env/environment.yml
@@ -6,7 +6,7 @@ dependencies:
   - pandas=2.2.0
   - jupyter=1.0.0
   - matplotlib=3.8.2
-  - ipykernel 6.29.0
+  - ipykernel=6.29.0
   - numpy=1.26.3
   - rasterio=1.3.9
   - seaborn=0.13.2
