### 2025-01-20 #7527 @atrisovic
Hello  @erexer and @yc-lin-geo,
I hope all is well. Just a gentle reminder about the JOSS review. It would be great to get an update on your progress with the revisions. Please let us know if you need any clarification or additional support. Thanks for your efforts!
Ana
