### 2025-01-28 #7527 @yc-lin-geo
I enjoy reading this manuscript. I have to state that I am not an expert on flooding but have just came across some of the articles in this area, so my comments should be regarded as a person with general background in climate science and software development.

Overall, I think this software is valueable for the community, and I would support this publication. But I just have one concern. Given the short paper length, I don't think people without enough background in flooding would be able to run it and use it for different purposes. For example, what is the difference between HAZUS and NACCS, why I should use any one of them, and for choosing either one product, will I face additional structural uncertainty for choosing depth-damage function? Providing more information about the assumption, basic theory of how these depth-damage function created and how do they can potentially impact on the modeling results would be something necessary. But otherwise I would be happy for this paper and software.
