### 2025-02-04 #7527 @erexer
> Hi [@erexer](https://github.com/erexer), thanks for your question. I led the software development, but the co-authors helped come up with the idea and supported design choices. [James](https://github.com/jdossgollin) reproduced the test cases in the `examples/` directory and critically reviewed the application of the package to test cases, which led to updates in both the package's codebase and reproduction instructions (the latter of which he made directly). I can add an Author Contributions statement to the README if that helps. Please let me know if you have other questions.

An Author Contribution statement would be great, thanks!
