### 2025-02-04 fa0ae60
add v0.1 contributions to readme
