### 2025-02-04 #7527 @abpoll
Thanks for the suggestion, @erexer - I updated the README. Should I release a new version or is it ok as-is?
