### 2025-02-04 #7527 @abpoll
Thank you @yc-lin-geo for your review. I tried to follow the concise formatting typical of JOSS papers. I completely agree with you that there are not many technical details in the paper. For that reason, my co-authors and I worked on a [technical documentation](https://github.com/abpoll/unsafe/blob/main/docs/v01.pdf) with the repository. In the long run I have hopes to use a better workflow to generate technical documentation (such as Sphinx), but we felt that for v0.1 of the repository, having the content in place was a good start.
