### 2025-02-04 abpoll/unsafe#10 @erexer
partial_data_example.ipynb questions
This is for the [JOSS Review](https://github.com/openjournals/joss-reviews/issues/7527).

For some reason in the [examples README](https://github.com/abpoll/unsafe/blob/main/examples/README.md#examples), when I click on the hyperlink [FEMA Flood Map Service Center: Search All Products landing page](https://msc.fema.gov/portal/advanceSearch), it brings me to a `403 Forbidden` page, but when I go to https://msc.fema.gov/portal/ and click on [Search All Products](https://msc.fema.gov/portal/advanceSearch) it brings me to the search page totally fine. Does this happen for you?

Some questions/comments:
- [x] pyyaml isn't listed in the dependencies but is required.
- [x] `partial_data_example.ipynb` typo: congig
- [x] `partial_data_example.ipynb`: When I run `ununzip.unzip_raw(FR, UNZIP_DIR)`, I get the error
```bash
BadZipFile                                Traceback (most recent call last)
Cell In[13], line 1
----> 1 ununzip.unzip_raw(FR, UNZIP_DIR)

File ~/<path>/unsafe/src/unsafe/unzip.py:103, in unzip_raw(fr, unzip_dir)
    100     subdir = filepath.split("/")[-1][:-4]
    101     out_filedir = join(out_filedir, subdir)
--> 103 with ZipFile(path, "r") as zip_ref:
    104     zip_ref.extractall(out_filedir)
    106 # TODO helpful log message

File ~/micromamba/envs/.venv_micromamba/lib/python3.9/zipfile.py:1266, in ZipFile.__init__(self, file, mode, compression, allowZip64, compresslevel, strict_timestamps)
   1264 try:
   1265     if mode == 'r':
-> 1266         self._RealGetContents()
   1267     elif mode in ('w', 'x'):
   1268         # set the modified flag so central directory gets written
   1269         # even if no files are added to the archive
   1270         self._didModify = True

File ~/micromamba/envs/.venv_micromamba/lib/python3.9/zipfile.py:1333, in ZipFile._RealGetContents(self)
   1331     raise BadZipFile("File is not a zip file")
   1332 if not endrec:
-> 1333     raise BadZipFile("File is not a zip file")
   1334 if self.debug > 1:
   1335     print(endrec)

BadZipFile: File is not a zip file
```
