### 2025-02-04 abpoll/unsafe#10 @abpoll
Hi @erexer , thanks for the very detailed comment.

> Does this happen for you?
Yes. That didn't happen before, but it did happen just now. I will update the instructions accordingly.

Re: your comments and questions:
1. pyyaml isn't listed in the dependencies but is required.

I added this to `pyproject.toml` for **UNSAFE** , since it's a required part of the workflow, and tested that the package shows up in the environment you install **UNSAFE** into. Note: the package was included in my other test environment based on the `environment.yml` file, but maybe that is unique to the version of mamba I'm using? Hopefully now that it is in `pyproject.toml` it will work for you.

2. Thanks for the note - I fixed the typo.

3. I have to look into this. I will comment back here after I dig around. I know that this week some people have had issues accessing the census.gov website so I wonder if *UNSAFE* needs a more reliable long-term way to download some of the data it uses...

I will push changes for 1 & 2 after a little more testing and get back to you with 3.
