### 2025-02-04 abpoll/unsafe#10 @abpoll
Ok, as I suspected, the issue is that now we can't access the US Census FTP to download data. This does limit one's ability to represent results over space in a nice way, but should not affect core functionality of **UNSAFE**. I'll have to update this too
