### 2025-02-04 e76421d
addressing first two issues with reproduction
