+++ pyproject.toml
@@ -11,7 +11,8 @@ dependencies= [
     "geopandas>=0.14",
     "rasterio>=1.3.9",
     "zipfile-deflate64>=0.2.0",
-    "pyarrow>=14.0.2"
+    "pyarrow>=14.0.2",
+    "pyyaml>=6.0.2"
 ]
 authors = [
   { name="Adam Pollack", email="adampollack25@gmail.com" },
