### 2025-02-04 abpoll/unsafe#10 @abpoll
Addressed 1 & 2 here: e76421da946e0f40ff1193678fcac78ee004f39c.

Re: 1 - `mamba install env/environment.yml` for the new test environment installed pyyaml, but it's in `pyproject.toml` now so I hope it works for you. Please let me know.

Re: 3 - I reproduced your error and I'm working on fixing it. Will keep you updated. From what I can tell, it might be as simple as avoiding the reference files (i.e., the Census data) in the `config_partial.yaml` file since it's not used in a meaningful way in this notebook to demonstrate core functionality. I am very helpful the US Census TIGER data will be back online soon. It's shocking, but I know a lot of other federal data was scrubbed from websites this week.
