### 2025-02-04 abpoll/unsafe#10 @abpoll
I found backed up Census data on the wayback machine. I updated the `config_partial.yaml` file with the appropriate URL endpoints and tested that the files download and unzip correctly. I removed the one of the previously downloaded files (the CDC SVI 2020 data) that is not necessary to demonstrate the core functionality since there is another social vulnerability dataset backed up and the core functions are demonstrated on that file. I did not update the config file or code in the `philadelphia_frd/` example because I am hopeful the Census data and CDC SVI data will be back online soon. But I can update it if required for the JOSS review.
