### 2025-02-14 abpoll/unsafe#10 @erexer
> For some reason in the [examples README](https://github.com/abpoll/unsafe/blob/main/examples/README.md#examples), when I click on the hyperlink [FEMA Flood Map Service Center: Search All Products landing page](https://msc.fema.gov/portal/advanceSearch), it brings me to a 403 Forbidden page, but when I go to https://msc.fema.gov/portal/ and click on [Search All Products](https://msc.fema.gov/portal/advanceSearch) it brings me to the search page totally fine. Does this happen for you?

This is still happening for me, I don't see that the README has changed?
