### 2025-02-14 aa6dba9
Update README.md
Change instructions & URL for downloading external data
