+++ examples/README.md
@@ -5,7 +5,7 @@ This subdirectory hosts two examples of worklows that make use of UNSAFE. The pu
 1) `../external/haz/dg_clipped.zip`, and
 2) `../external/vuln/ddfs.zip`
 
-Because these two datasets are provided for you, you can run all cells in the `phil_frd_partial/notebooks/partial_data_example.ipynb` notebook. The hazard data in the dg_clipped.zip directory is a clipped depth grid from the full data example included in the `philadelphia_frd/` example. This example includes the depth-damage functions .zip directory, but the hazard data is too large to upload to GitHub. To work through the full data example, you will need to upload the FEMA flood risk database for Philadelphia into `../external/haz/`. You can find this dataset by going go the [FEMA Flood Map Service Center: Search All Products landing page](https://msc.fema.gov/portal/advanceSearch) and searching for Product ID FRD_02040202_PA_GeoTIFFs. Alternatively, you can search under "Jurisdiction" for PENNSYLVANIA -> PHILADELPHIA COUNTY -> PHILADELPHIA COUNTY ALL JURISDICTIONS and then click "Search." As of April of 2024 you will see directories named "Effective Products," "Preliminary Products," "Pending Product," "Historic Products," and "Flood Risk Products." You can click on Flood Risk Products -> Flood Risk Database and then download the GeoTIFFs file. The one used in the full data example was posted on 08/01/2016 and is 2190MB. By the time you are reading this, it's possible the data has been updated, or the website looks different, and we can explore options to use a new FEMA flood risk database product if it's available or check permission with FEMA for us to upload the version we used for our example on a repository like Zenodo.  
+Because these two datasets are provided for you, you can run all cells in the `phil_frd_partial/notebooks/partial_data_example.ipynb` notebook. The hazard data in the dg_clipped.zip directory is a clipped depth grid from the full data example included in the `philadelphia_frd/` example. This example includes the depth-damage functions .zip directory, but the hazard data is too large to upload to GitHub. To work through the full data example, you will need to upload the FEMA flood risk database for Philadelphia into `../external/haz/`. You can find this dataset by going go the [FEMA Flood Map Service Center: Welcome landing page](https://msc.fema.gov/portal/home), clicking on Search All Products (highlighted in blue just below "Looking for more than just a current flood map?" and searching for Product ID FRD_02040202_PA_GeoTIFFs. Alternatively, you can search under "Jurisdiction" for PENNSYLVANIA -> PHILADELPHIA COUNTY -> PHILADELPHIA COUNTY ALL JURISDICTIONS and then click "Search." As of April of 2024 you will see directories named "Effective Products," "Preliminary Products," "Pending Product," "Historic Products," and "Flood Risk Products." You can click on Flood Risk Products -> Flood Risk Database and then download the GeoTIFFs file. The one used in the full data example was posted on 08/01/2016 and is 2190MB. By the time you are reading this, it's possible the data has been updated, or the website looks different, and we can explore options to use a new FEMA flood risk database product if it's available or check permission with FEMA for us to upload the version we used for our example on a repository like Zenodo.  
 
 You can also test out UNSAFE outside of these examples with your own hazard data in the same county or a different county. To do this, you will need to upload depth-damage functions into the `../data/raw/external/` subdirectory of your project, which you can download from this [Zenodo repository](https://zenodo.org/records/10027236).
 
