### 2025-02-14 abpoll/unsafe#10 @abpoll
My apologies. I just committed the changes: https://github.com/abpoll/unsafe/commit/aa6dba9a2343910a925295dd16ce954ca047a572
