### 2025-02-14 abpoll/unsafe#10 @erexer
Unsure if this is related to 3, but I'm unable to access this site: https://static-data-screeningtool.geoplatform.gov/data-versions/1.0/data/score/downloadable/ and this is failing:

```python
undown.download_raw(DOWNLOAD, wcard_dict,
                    FR, API_EXT)
```
```
Downloaded from: https://nsi.sec.usace.army.mil/nsiapi/structures?fips=42101
Downloaded from: https://hazards.fema.gov/nfhlv2/output/County/420757_20230701.zip
Downloaded from: https://web.archive.org/web/20231130105526/https://www2.census.gov/geo/tiger/TIGER2022/TRACT/tl_2022_42_tract.zip
Downloaded from: https://web.archive.org/web/20240604063409/https://www2.census.gov/geo/tiger/TIGER2022/BG/tl_2022_42_bg.zip
Downloaded from: https://web.archive.org/web/20240608162400/https://www2.census.gov/geo/tiger/TIGER2022/TABBLOCK20/tl_2022_42_tabblock20.zip
SSLError                                  Traceback (most recent call last)
File ~/micromamba/envs/unsafe/lib/python3.12/site-packages/urllib3/connectionpool.py:464, in HTTPConnectionPool._make_request(self, conn, method, url, body, headers, retries, timeout, chunked, response_conn, preload_content, decode_content, enforce_content_length)

...

SSLError: HTTPSConnectionPool(host='static-data-screeningtool.geoplatform.gov', port=443): Max retries exceeded with url: /data-versions/1.0/data/score/downloadable/1.0-communities.csv (Caused by SSLError(SSLError(1, '[SSL: SSLV3_ALERT_HANDSHAKE_FAILURE] ssl/tls alert handshake failure (_ssl.c:1000)')))
```
