### 2025-02-14 abpoll/unsafe#10 @erexer
> My apologies. I just committed the changes: https://github.com/abpoll/unsafe/commit/aa6dba9a2343910a925295dd16ce954ca047a572

Weirdly enough, I'm still getting this error with the url https://msc.fema.gov/portal/home when I follow it from the README, if I load the same url from a new tab it works? Might be an issue with my machine though.
