### 2025-02-14 83721fe
update config files with persistent link to cejst data
