+++ examples/philadelphia_frd/config/config.yaml
@@ -53,7 +53,7 @@ download:
       vuln:
         social:
           # CEJST and CDC SVI
-          cejst: "https://static-data-screeningtool.geoplatform.gov/data-versions/1.0/data/score/downloadable/1.0-communities.csv"
+          cejst: "https://dblew8dgr6ajz.cloudfront.net/data-versions/2.0/data/score/downloadable/2.0-communities.csv"
           svi: "https://svi.cdc.gov/Documents/Data/2020/csv/states/SVI_2020_US.csv"
       ref:
         # County and zip code tabulation area spatial data available for the whole country
