### 2025-02-14 abpoll/unsafe#10 @abpoll
> Unsure if this is related to 3, but I'm unable to access this site: https://static-data-screeningtool.geoplatform.gov/data-versions/1.0/data/score/downloadable/ and this is failing:
>
> undown.download_raw(DOWNLOAD, wcard_dict,
>                     FR, API_EXT)

I'm sorry about all of these data issues. It turns out that the white house took down the climate and economic justice screening tool data as well. I found a website committed to maintaining and preserving this kind of data long term (<https://screening-tools.com/>) and replaced both `config.yaml` and `config_partial.yaml` with these links to support the reproducibility testing.

I tested that the data downloads and unzips correctly and that the rest of the notebook works as expected. I'll leave the issue open.

> Weirdly enough, I'm still getting this error with the url https://msc.fema.gov/portal/home when I follow it from the README, if I load the same url from a new tab it works? Might be an issue with my machine though.

Sorry, I can't reproduce this error. Would you like me to add an instruction to the README about the issue? Happy to make it more user friendly.
