### 2025-03-04 c59cf95
use ddf_type instead of ddf_id as more generalizable feature for linking structures and ddfs
