### 2025-03-04 7767c16
more accessible readme
