+++ README.md
@@ -1,54 +1,80 @@
 # Uncertain Structure and Fragility Ensemble (**UNSAFE**) framework for property-level flood risk estimation
 
-This repository hosts the latest **UNSAFE** version.
-It also includes example code that demonstrates a workflow for using **UNSAFE** with flood inundation model outputs to estimate property-level flood risks.
+[![Version](https://img.shields.io/badge/version-0.1-blue.svg)](https://github.com/abpoll/unsafe)
+[![License](https://img.shields.io/badge/License-BSD--2--Clause-green.svg)](LICENSE)
+[![Maintenance](https://img.shields.io/badge/Maintenance-Active-brightgreen.svg)](Maintenance)
 
-## Statement of need
+**UNSAFE** is an open-source framework for estimating property-level flood risk that explicitly accounts for uncertainties in exposure and vulnerability.
 
-Flooding is a frequent, widespread, and damaging natural hazard in the United States.
-Property-level economic flood risk estimates are increasingly used in research and practice to help inform management practices and policies. Economic flood risk is estimated as a function of hazard (_i.e._, the features of a flood over space and time), exposure (_i.e._, the assets that experience inundation from a flood), and vulnerability (_i.e._, the susceptibility of exposed assets to damage for a set of flood features).
-
-It is common for property-level economic flood risk assessments to overlook uncertainty in these inputs.
-When uncertainty is incorporated, it is more common to account for uncertainty in flood hazard than in exposure or vulnerability.
-While it is a common finding that flood risk estimates are most sensitive to uncertainty in flood hazard estimates, overlooking uncertainty in exposure and vulnerability can bias risk estimates.
-
-The Uncertain Structure and Fragility Ensemble (**UNSAFE**) framework aims to fill the need for a published, free, and open-source tool for representing exposure and vulnerability under uncertainty for flood-risk estimation in the U.S.
-We invite others to contribute to this project to help standardize best practices in the estimation of flood risk under uncertainty, improve reusability and efficiency, expand functionality for more use-cases, and maintain a state-of-the-art risk estimation codebase that is free and usable by any interested party.
+## Table of Contents
+- [Overview](#overview)
+- [Statement of Need](#statement-of-need)
+- [Installation](#installation)
+- [Examples](#examples)
+- [Documentation](#documentation)
+- [Contributiong](#contributions)
+- [License](#license)
+- [Citation](#citation)
+- [Acknowledgements](#acknowledgements)
 
 ## Overview
+The Uncertain Structure and Fragility Ensemble (**UNSAFE**) framework enhances a property-level risk assessment framework common in academic research and practice (_e.g._, [Federal Emergency Management Agency (FEMA) loss avoidance studies](https://www.fema.gov/grants/mitigation/loss-avoidance-studies), [United States Army Corps of Engineers (USACE) feasibility studies](https://www.nad.usace.army.mil/Portals/40/docs/NACCS/10A_PhysicalDepthDmgFxSummary_26Jan2015.pdf)). At a high-level, **UNSAFE**:
 
-**UNSAFE** modifies a property-level risk assessment framework common in academic research and practice (_e.g._, [Federal Emergency Management Agency (FEMA) loss avoidance studies](https://www.fema.gov/grants/mitigation/loss-avoidance-studies), [United States Army Corps of Engineers (USACE) feasibility studies](https://www.nad.usace.army.mil/Portals/40/docs/NACCS/10A_PhysicalDepthDmgFxSummary_26Jan2015.pdf)).
[…]
