+++ CITATION.cff
@@ -1,0 +1,43 @@
+# This CITATION.cff file was generated with cffinit.
+# Visit https://bit.ly/cffinit to generate yours today!
+
+cff-version: 1.2.0
+title: UNSAFE
+message: >-
+  If you use this software, please cite it using the
+  metadata from this file.
+type: software
+authors:
+  - given-names: Adam
+    family-names: Pollack
+    orcid: 'https://orcid.org/0000-0001-6642-0591'
+    affiliation: Dartmouth College
+  - given-names: James
+    family-names: Doss-Gollin
+    affiliation: Rice University
+    orcid: 'https://orcid.org/0000-0002-3428-2224'
+  - given-names: Vivek
+    family-names: Srikrishnan
+    affiliation: Cornell University
+    orcid: 'https://orcid.org/0000-0003-0049-3805'
+  - given-names: Klaus
+    family-names: Keller
+    affiliation: Dartmouth College
+    orcid: 'https://orcid.org/0000-0002-5451-8687'
+identifiers:
+  - type: doi
+    value: 10.31219/osf.io/jb9ta
+repository-code: 'https://github.com/abpoll/unsafe'
+abstract: >-
+  A software package for adding parametric uncertainty to
+  the national structure inventory and estimating flood
+  losses with uncertain depth-damage relationships.
+keywords:
+  - flood risk
+  - uncertainty
+  - monte carlo
+  - depth-damage function
+license: BSD-2-Clause
+commit: 0f36bb02f5e30ad6d7302bef956a6ce5e2a92b63
+version: v0.1
+date-released: '2024-05-14'
