### 2025-03-05 1af419c
improved ddf functions documentation
