+++ src/unsafe/ddfs.py
@@ -1,3 +1,12 @@
+"""
+UNSAFE: Depth-Damage Functions
+
+This module provides functions for processing and applying 
+depth-damage functions to estimate flood damages. It supports multiple
+DDF sources and uncertainty quantification.
+
+"""
+
 # Packages
 import pandas as pd
 import numpy as np
@@ -7,16 +16,46 @@ import json
 from unsafe.const import *
 from unsafe.files import *
 
-"""
-Helpful functions for processing depth-damage functions
-that are stored in different data structures
-"""
 
-
-# HAZUS and USACE DDFs can be
-# formatted in tidy df format
-# this way
 def tidy_ddfs(raw_ddf, idvars):
+    """
+    Convert depth-damage function data from wide to tidy (long) format.
+    
+    This function transforms depth-damage function data from wide format (where
+    depths are columns) to long format (where depths are rows). It also processes
+    depth values, converting them from string format (e.g., "1ft", "-1m") to 
+    numeric values in feet, and normalizes damage percentages to proportions.
+    
+    Parameters
+    ----------
+    raw_ddf : pandas.DataFrame
+        Raw depth-damage function data in wide format, where:
+        - Columns include identifier variables and depth columns
+        - Values represent damage percentages (0-100)
+    
+    idvars : list
+        List of column names to use as identifier variables (not to be melted).
+        Examples: ['ddf_id', 'occ_type', 'dam_cat']
+    
+    Returns
+    -------
+    pandas.DataFrame
+        Tidy (long-format) dataframe with columns:
+        - All columns from idvars
+        - depth_str: Original depth string (e.g., "1ft")
+        - pct_dam: Original damage percentage (0-100)
+        - depth_ft: Depth in feet as float
+        - rel_dam: Relative damage as proportion (0-1)
+    
+    Notes
+    -----
+    Wide format is the standard way HAZUS DDFs are available, so other DDFs
+    tend to have the same raw data structure so they can be processed similarly.
+    Depth strings with 'm' suffix are treated as negative depths (e.g., "-1m").
+    All depths are converted to feet.
+    """
+    print("Converting depth-damage functions to tidy (long) format...")
+
     # For HAZUS we have occtype and fld_zone
     # For NACCS we have Occupancy and DamageCategory
     ddf_melt = raw_ddf.melt(id_vars=idvars, var_name="depth_str", value_name="pct_dam")
@@ -45,6 +84,34 @@ def tidy_ddfs(raw_ddf, idvars):
 
 
 def ddf_max_depth_dict(tidy_ddf, dam_col):
+    """
+    Create a dictionary mapping DDF types to damage parameters at maximum depths
+    of the DDF data.
+    
+    This function identifies the maximum depth for each depth-damage function (DDF) type
+    and creates a lookup dictionary that maps each DDF type to its damage parameters
+    at that maximum depth. This is used to handle cases where flood depths exceed
+    the maximum depth defined in the DDF.
+    
+    Parameters
+    ----------
+    tidy_ddf : pandas.DataFrame
+        A tidy (long-format) dataframe containing DDF data with columns:
+        - ddf_type: Identifier for the DDF type
+        - depth_ft: Depth values in feet
+        - dam_col: Column name containing damage parameters
+    
+    dam_col : str
+        Name of the column containing damage parameters (e.g., 'params')
+    
+    Returns
+    -------
+    dict
+        Dictionary mapping DDF type identifiers to their damage parameters
+        at maximum depth. Format: {ddf_type: damage_params}
+    """
[…]
