### 2025-03-05 f3c9d58
experimenting with generalizing the codebase for multiple structure inventories
