+++ src/unsafe/download.py
@@ -6,28 +6,12 @@ import yaml
 from yaml.loader import SafeLoader
 import requests
 import json
-from unsafe.files import *
-from unsafe.const import *
+import unsafe.files as unfile
+import unsafe.const as unconst 
 
 """
 Define our utils
 """
-
-
-# The fill_wcard function
-# For any URL/API endpoint or directory, we want to replace
-# wildcard terms (FIPS, STATE, STATE_ALPHA)
-# with the appropriate value from the FIPS, STATE, or
-# STATE_ALPHA config values
-def fill_wcard(endpoint, wcard_dict):
-    # Get a list of all the wildcards we need to replace for this endpoint
-    wildcards = [wcard for wcard in wcard_dict.keys() if wcard in endpoint]
-    # Replace the wildcard with the value stored in a wildcard dictionary
-    for wildcard in wildcards:
-        endpoint = endpoint.replace(wildcard, wcard_dict[wildcard])
-
-    return endpoint
-
 
 # The get_dir helper function
 # For a list of string tokens, we
@@ -98,48 +82,6 @@ def process_file(file):
 
     return str_tokens, endpoint
 
-
-# The dwnld_out_files function
-# For each URL/API endpoint, we want to return the output
-# version of that file
-# This is the function we need for the download_unzip_data
-# rule output
-# TODO this works now, but when we have multiple
-# counties and states this will need to keep
-# the wildcards from the file passed
-# to process_file and add it to the directory
-# structure and/or filename
-def dwnld_out_files(files):
-    # For each file that needs to be downloaded
-    # Get the filepath
-    # Return the list of these out files
-    out_list = []
-    for file in files.itertuples():
-        str_tokens, endpoint = process_file(file)
-
-        # Helper function to return
-        # our filepath from our str_tokens
-        # The last token is our filename
-        # Everything after the second token and before this
-        # are file directories
-        # For API endpoints, we need to access the "ext" element
-        # to get the filename (urls include the file ext)
-        # We use the exts dict for this
-        filepath = get_dir(str_tokens, endpoint)
-
-        # Replace the wildcard text (i.e. FIPS, STATEABBR)
-        # with the text "unit" so that the download_data
-        # rule only has one wildcard for each
-        # output file
-        out_dict = {"{FIPS}": "{unit}", "{STATEABBR}": "{unit}", "{NATION}": "{unit}"}
-        filepath = fill_wcard(filepath, out_dict)
-
-        # Add this filepath to our out_list
-        out_list.append(filepath)
-
-    return out_list
-
-
 # The download_url helper function
 def download_url(url, save_path, chunk_size=128):
     r = requests.get(url, stream=True)
@@ -174,12 +116,12 @@ def download_raw(files, wcard_dict, fr, api_ext):
         # Get the out filepath
         # "Clean" it with the wcard_dict
         out_filepath = get_dir(str_tokens, endpoint, fr, api_ext)
-        out_filepath = fill_wcard(out_filepath, wcard_dict)
+        out_filepath = unfile.fill_wcard(out_filepath, wcard_dict)
         # "Clean" the endpoint with the wcard_dict
-        endpoint = fill_wcard(endpoint, wcard_dict)
+        endpoint = unfile.fill_wcard(endpoint, wcard_dict)
 
         # Make sure we can write out data to this filepath
-        prepare_saving(out_filepath)
+        unfile.prepare_saving(out_filepath)
 
         # Download data with api or url call
         if str_tokens[1] == "api":
