### 2025-03-18 6f81d58
major revisions to naccs damage procedure - use res1 basement ddfs for apartments with basements for more accurate benchmark
