+++ src/unsafe/ensemble.py
@@ -8,9 +8,8 @@ import geopandas as gpd
 import pandas as pd
 import numpy as np
 from unsafe.files import *
-from unsafe.const import *
 from unsafe.ddfs import *
-
+import unsafe.const as unconst
 
 def get_base_df(fips, exp_dir_i):
     """
@@ -105,6 +104,358 @@ def get_base_df(fips, exp_dir_i):
 
     return full_df
 
+def get_loss_ensemble(
+    structures_df,
+    depths_df,
+    config=None,
+    vuln_dir=None,
+    random_seed=None
+):
+    
+    """
+    Generate an ensemble of structure realizations with uncertainty and calculate damages.
+    
+    This function creates multiple realizations of structures by introducing
+    uncertainty in selected characteristics and calculating damages.
+    
+    Parameters
+    ----------
+    structures_df : pandas.DataFrame or geopandas.GeoDataFrame
+        DataFrame containing all structures with their characteristics.
+        Must have a unique identifier column (specified in config['id_col']).
+    
+    depths_df : pandas.DataFrame
+        DataFrame containing depth values for structures.
+        Index or id_column should match structures_df's id_column.
+        Only structures with non-zero depths will be used for ensemble generation.
+    
+    vuln_dir : str or Path
+        Path to directory containing processed DDF files.
+    
+    config : dict, optional
+        Configuration dictionary with the following keys:
+        - 'n_sow': Number of ensemble members (default: 1000)
+        - 'struct_list': List of characteristics to vary (e.g., ['ffe', 'val_struct'])
+        - 'ffe_dict': Dict mapping foundation types to FFE distribution parameters
+        - 'coef_var': Float for structure value uncertainty (default: 0.2)
+        - 'ref_col': Column name for spatial reference (default: 'tract_id')
+        - 'id_col': Column name for structure identifier (default: 'fd_id')
+        - 'ddfs': List of depth-damage function types to use (e.g., ['hazus', 'naccs'])
+        - 'base_adj': Boolean for starting basement flooding from bottom of basement
+        - 'found_param': Prop. of structures in each ref_id area with various foundation types
+        - 'stories_param': Prop. of structures in each ref_id area with various stories
+    
+    random_seed : int, optional
+        Random seed for reproducibility.
+    
+    Returns
+    -------
+    pandas.DataFrame
+        DataFrame containing the ensemble with n_sow realizations of each structure.
+        Includes columns for each uncertain characteristic and damage estimates.
+    
+    Notes
+    -----
+    The function follows this workflow:
+    1. Generate reference distributions from the full structures dataset
+    2. Filter structures to those with non-zero depths (if depths_df provided)
+    3. Generate ensemble members for the filtered structures
+    4. Calculate damages using the specified DDFs
+    """
+    # If required columns are not all present, throw exception
+    req_cols = ['tract_id', 'val_struct', 'num_story',
+                'found_type', 'occtype']
+    if not all([x in structures_df.columns for x in req_cols]):
+        raise KeyError('Structures DataFrame missing required column: ' + str(req_cols))
+
+    # Set random seed if provided
+    if random_seed is not None:
+        np.random.seed(random_seed)
+    
+    # Set up random number generator
+    rng = np.random.default_rng(random_seed)
+    
+    # Set default configuration if not provided
+    if config is None:
+        config = {}
+    
+    # Extract configuration parameters with defaults
+    n_sow = config.get('n_sow', 1000)
+    struct_list = config.get('struct_list', ['val_struct'])
[…]
