+++ src/unsafe/ddfs.py
@@ -212,11 +212,15 @@ def process_naccs(vuln_dir_uz, vuln_dir_i):
     for ddf_type, df in naccs_piv.groupby(["ddf_id", "occ_type"]):
         print(f"  Interpolating depths for {ddf_type}...")
         # This creates the duplicate rows
-        ddf_int = df.loc[np.repeat(df.index, 10)].reset_index(drop=True)
+        # from min to max depth sampled at .1 feet
+        new_depths = np.arange(df['depth_ft'].min(),
+                               df['depth_ft'].max() + .1,
+                               .1).round(1)
+        new_depths = pd.Series(new_depths, name='depth_ft')
+        ddf_int = df.merge(new_depths, on='depth_ft', how='right')
         # Now we have to make them nulls by finding
         # the "original" indexed rows
         float_cols = ["depth_ft", "ML", "Max", "Min"]
-        ddf_int.loc[ddf_int.index % 10 != 0, float_cols] = np.nan
         # Now we interpolate (again, on floats)
         ddf_int_floats = ddf_int[float_cols].interpolate().round(2)
         # Add in our ddf type col
@@ -485,14 +489,33 @@ def process_hazus(vuln_dir_uz, vuln_dir_i, unif_unc=0.3):
     print("HAZUS DDFs processed")
 
 
-def est_naccs_loss(bld_types, depths, ddfs, MAX_DICT):
+def est_naccs_loss(ddf_types, depths, ffes, ddfs, MAX_DICT, base_adj):
     # We are going to use a random number generator
     rng = np.random.default_rng()
 
     # Combine building types and depths on index
-    bld_depths = pd.concat([bld_types, pd.Series(depths)], axis=1)
+    bld_depths = pd.concat([pd.Series(ddf_types),
+                            pd.Series(depths),
+                            pd.Series(ffes)], axis=1)
     # Rename columns to correspond to each series
-    bld_depths.columns = ["ddf_type", "depth_ft"]
+    bld_depths.columns = ["ddf_type", "depth", "ffe"]
+
+    # Get depth relative to first floor elevation, considering basement
+    # adjustment
+    BSMT_HGT = 9
+    wb_obs = bld_depths['ddf_type'].str[2:4] == 'WB'
+    bld_depths.loc[~wb_obs, 'depth_ft'] = (bld_depths.loc[~wb_obs, 'depth'] 
+                                            - bld_depths.loc[~wb_obs, 'ffe'])
+    if base_adj:
+        bld_depths.loc[wb_obs, 'depth_ft'] = (bld_depths.loc[wb_obs, 'depth']
+                                               - BSMT_HGT)
+    else:
+        bld_depths.loc[wb_obs, 'depth_ft'] = (bld_depths.loc[wb_obs, 'depth']
+                                               - bld_depths.loc[wb_obs, 'ffe'])
+
+    # Round to the precision of our depth-damage functions
+    bld_depths['depth_ft'] = bld_depths['depth_ft'].round(1).copy()
+
     # Merge bld_type/depths with the ddfs to get params linked up
     # Need to create columns for the merge
     # We use the number 100, since this is our floating point precision
@@ -507,6 +530,9 @@ def est_naccs_loss(bld_types, depths, ddfs, MAX_DICT):
     )
     # Drop the merge column
     loss_prep = loss_prep.drop(columns="merge")
+    # Update depth_ft to null if below basement depth
+    below_base = loss_prep['depth_ft'] < -BSMT_HGT
+    loss_prep.loc[below_base, 'depth_ft'] = np.nan
     # Helpful to have a mask for where there are no flood depths
[…]
