### 2025-05-02 b7f753a
some endpoints dont have wildcard tokens
