### 2025-05-02 8d8a9db
filter depths before converting to feet
