### 2025-05-27 #7527 @abpoll
Please let me know if I can help, too. One thing I can do to make the code review easier is to create an example that doesn't download any data from online, which was the source of earlier issues. Downloading data is not really a core function of the workflow (just something that was -- ironically -- supposed to help make things easier/more reproducible). Thanks everyone!
