### 2025-05-28 1cfd28b
updating examples to work with latest enhancements and updating bad data links
