+++ examples/README.md
@@ -5,9 +5,9 @@ This subdirectory hosts two examples of worklows that make use of UNSAFE. The pu
 1) `../external/haz/dg_clipped.zip`, and
 2) `../external/vuln/ddfs.zip`
 
-Because these two datasets are provided for you, you can run all cells in the `phil_frd_partial/notebooks/partial_data_example.ipynb` notebook. The hazard data in the dg_clipped.zip directory is a clipped depth grid from the full data example included in the `philadelphia_frd/` example. This example includes the depth-damage functions .zip directory, but the hazard data is too large to upload to GitHub. To work through the full data example, you will need to upload the FEMA flood risk database for Philadelphia into `../external/haz/`. You can find this dataset by going go the [FEMA Flood Map Service Center: Welcome landing page](https://msc.fema.gov/portal/home), clicking on Search All Products (highlighted in blue just below "Looking for more than just a current flood map?" and searching for Product ID FRD_02040202_PA_GeoTIFFs. Alternatively, you can search under "Jurisdiction" for PENNSYLVANIA -> PHILADELPHIA COUNTY -> PHILADELPHIA COUNTY ALL JURISDICTIONS and then click "Search." As of April of 2024 you will see directories named "Effective Products," "Preliminary Products," "Pending Product," "Historic Products," and "Flood Risk Products." You can click on Flood Risk Products -> Flood Risk Database and then download the GeoTIFFs file. The one used in the full data example was posted on 08/01/2016 and is 2190MB. By the time you are reading this, it's possible the data has been updated, or the website looks different, and we can explore options to use a new FEMA flood risk database product if it's available or check permission with FEMA for us to upload the version we used for our example on a repository like Zenodo.  
+Because these two datasets are provided for you, you can run all cells in the `phil_frd_partial/notebooks/partial_data_example.ipynb` notebook. The hazard data in the dg_clipped.zip directory is a clipped depth grid from the full data example included in the `philadelphia_frd/` example. This example includes the depth-damage functions .zip directory, but the hazard data is too large to upload to GitHub. To work through the full data example, you may access the hazard data [here](https://doi.org/10.5281/zenodo.15538686). Please download this data and place the `dg.zip` directory in `data/raw/external/haz/` (you will need to create the `haz` subdirectory). You do not have to unzip this, as **UNSAFE** will do this for you. 
 
-You can also test out UNSAFE outside of these examples with your own hazard data in the same county or a different county. To do this, you will need to upload depth-damage functions into the `../data/raw/external/` subdirectory of your project, which you can download from this [Zenodo repository](https://zenodo.org/records/10027236).
+You can also test out UNSAFE outside of these examples with your own hazard data in the same county or a different county. To do this, you will need to upload depth-damage functions into a `../data/raw/external/vuln/` subdirectory of your project, which you can download from this [Zenodo repository](https://zenodo.org/records/10027236). We recommend placing your hazard data in `data/raw/external/haz/` as **UNSAFE** uses parts of the subdirectory paths when unzipping data to make it easier to find. 
 
 ## Getting ready to run the examples
 To run these examples, you will need to have have [conda](https://docs.conda.io/en/latest/) or [mamba](https://mamba.readthedocs.io/en/latest/) installed. Before you head over to the examples and launch the Jupyter notebooks, you will need to follow these steps:
