+++ examples/env/environment.yml
@@ -11,3 +11,4 @@ dependencies:
   - rasterio=1.3.9
   - seaborn=0.13.2
   - pyarrow=14.0.2
+  - rioxarray>=0.17
