### 2025-05-28 f12a58e
need to update hazus functions to work with other enhancements
