### 2025-05-28 9bf0de7
rm outdated ensemble generation function
