+++ src/unsafe/ensemble.py
@@ -465,268 +465,6 @@ def get_loss_ensemble(
 
     return final_ens_df
 
-def generate_ensemble(
-    nsi_sub, base_df, ddf_list, struct_list, n_sow, ffe_dict, coef_variation, vuln_dir_i
-):
-    """
-    Return a dataframe of length number of structures * N_SOW
-    This dataframe has N_SOW realizations of the uncertain
-    structures specified in struct_list and corresponding
-    estimates of losses. base_df is subset
-    to properties with non-zero hazard and nsi_sub is not. We
-    learn from nsi_sub to sample certain characteristics. We also
-    pass in constants that are defined from config.yaml
-
-    nsi_sub: DataFrame of the relevant characteristics from the NSI
-    for all structures like the ones we will estimate losses for
-    base_df: DataFrame of the relevant characteristics from the NSI
-    subset to the structures we will estimate losses for
-    ddf_list: list of ddfs as strings we will estimate losses from
-    struct_list: list of characteristics as strings we wish to sample
-    from
-    n_sow: int, number of ensemble members per structure
-    ffe_dict: dict, maps foundation types to first floor elevation
-    distribution
-    stry_dict: dict, maps tract ids to number of stories distribution
-    fnd_dict: dict, maps tract ids to foundation type distribution
-    coef_variation: float, what we scale the NSI value point estimate
-    by to get the standard deviation of the structure value distribution
-    """
-    # We need a randon number generator
-    rng = np.random.default_rng()
-
-    # TODO we might want to allow the user to specify
-    # whether they'd like to use a different reference id
-    # than census tract. Maybe this can be specified in
-    # the config?
-    STRUCT_REF = "tract_id"
-    # We create the tract_id column from the cbfips column
-    # provided in the nsi data
-    nsi_sub["tract_id"] = nsi_sub["cbfips"].str[:11]
-
-    # We need a dataframe of the right size and index
-    ens_df = base_df.loc[np.repeat(base_df.index, n_sow)].reset_index(drop=True)
-    print("Created Index for Ensemble")
-
-    # First, we will subset nsi_sub to a dataframe that
-    # corresponds to all entries in the same census tract
-    # as those in base_df. This approach assumes there is no correlation of
-    # the foundation type proportions and number of stories
-    # with features that vary within a census tract.
-    struct_tot = nsi_sub[nsi_sub[STRUCT_REF].isin(base_df[STRUCT_REF])]
-    # Also make sure the foundation type is one of C, S, or B
-    # TODO this should be updated to link more foundation types
-    # and damage functions
-    struct_tot = struct_tot[struct_tot["found_type"].isin(["B", "C", "S"])]
-
-    # If a variable is in struct_list, we will sample from a distribution
-    # If not, we will use the single value from base_df
-
-    if "val_struct" in struct_list:
-        # Draw from the structure value distribution for each property
-        # normal(val_struct, val_struct*CF_DET) where these are array_like
-        # Using 1 as an artificial, arbitrary lower bound on value
-        # Very low probability of getting a negative number but we cannot
-        # allow that because you cannot have negative risk
-        vals = rng.normal(ens_df["val_struct"], ens_df["val_struct"] * coef_variation)
-        vals[vals < 1] = 1
-    else:
-        vals = ens_df["val_struct"].copy()
-
-    print("Draw values")
-
-    if "stories" in struct_list:
-        # Determine the #stories distribution
-        # Get the total number of structures w/ number of stories
-        # in each block gruop
[…]
