### 2025-05-28 abpoll/unsafe#10 @abpoll
FYI I uploaded the depth grids to a Zenodo repository here because the FEMA webpage apparently no longer hosts any of the required data: <https://zenodo.org/records/10027236>. I also updated all of the download URLs in the config.yaml, which I tested in a new branch and could merge in to main.
