### 2025-06-10 079e6b6
better subset logic for nsi
