### 2025-06-11 86871b8
benchmark loss depth filter applied before conversion
