### 2025-06-23 #7527 @crvernon
@atrisovic you could revisit two where were too busy at the time of initial pre-review but who may be freed up now:

- JannisHoch
- jlarsen-usgs

Here are several other options as well from the UQ side of things:

- shulele
- baxmittens
- EwoutH
