### 2025-09-11 abpoll/unsafe#13 @erexer
Docs questions / feedback
This is for the [JOSS Review](https://github.com/openjournals/joss-reviews/issues/7527).

It's looking great! There's a couple things that JOSS requires that you need

- [x] `Automated Tests: Are there automated tests or manual steps described so that the functionality of the software can be verified?`
- Tests would be great, but I see that you have an [example output for the partial example](https://github.com/abpoll/unsafe/blob/main/examples/phil_frd_partial/fig/example_output.png), could you make it clear that you can use that to  verify the outputs. I see that there's a [link](https://github.com/abpoll/unsafe/tree/main/examples) for the output of the full data example, but it appears to be broken.
- [x] > `Community guidelines: Are there clear guidelines for third parties wishing to 1. Contribute to the software 2. Report issues or problems with the software 3. Seek support`
- Can you add a little more specific community guidelines? GitHub has a [guide](https://docs.github.com/en/communities/setting-up-your-project-for-healthy-contributions/setting-guidelines-for-repository-contributors) where they recommend modifying something like the [GitHub](https://github.com/github/docs/blob/d690e8a9028130f29cd181570d296263945d1f12/.github/CONTRIBUTING.md) or the [Ruby on Rails](https://github.com/rails/rails/blob/main/CONTRIBUTING.md) contribution guides.
- [x] In the [paper description of figure 1](https://github.com/abpoll/unsafe/blob/main/joss_submission/paper.md#summary) the example Jupyter notebook link is broken.

I'd also suggest (not required) linking to the docs in the README.
