### 2025-09-11 fa9d184
clarify instructions for checking output in readme of examples/ directory
