### 2025-09-11 0d3bee5
added contribution file and made more references to the outputs for examples for verification
