+++ CONTRIBUTING.MD
@@ -1,0 +1,25 @@
+# Thank you for contributing to `UNSAFE`!
+Thank you for investing your time in contributing to our project! We will recognize any contribution you make to the project. 
+
+In this guide, we provide an overview of the contribution workflow from opening an issue, creating a PR, reviewing, and merging the PR.
+
+## Getting started
+Please check out the [Installation](https://github.com/abpoll/unsafe?tab=readme-ov-file#installation) instructions on our README. 
+
+Run the [partial data example](https://github.com/abpoll/unsafe/tree/main/examples/phil_frd_partial) and verify your outputs [here](https://html-preview.github.io/?url=https://github.com/abpoll/unsafe/blob/main/examples/phil_frd_partial/notebooks/partial_data_example.html). 
+
+Run the [full data example](https://github.com/abpoll/unsafe/tree/main/examples/philadelphia_frd) and verify your outputs [here](https://html-preview.github.io/?url=https://github.com/abpoll/unsafe/blob/main/examples/philadelphia_frd/notebooks/full_data_example.html).
+
+## Issues
+### Create an issue
+If you spot a problem with the codebase or documentation, [search if an issue already exists](https://docs.github.com/en/github/searching-for-information-on-github/searching-on-github/searching-issues-and-pull-requests#search-by-the-title-body-or-comments). If a related issue doesn't exist, you can open a new issue using a relevant [issue form](https://github.com/abpoll/unsafe/issues/new).
+
+### Solve an issue
+Scan through our [existing issues](https://github.com/abpoll/unsafe/issues/new) to find one that interests you. You can narrow down the search using `labels` as filters. If you are new to contributing, you may want to check out the labels issued as ["good first issue"](https://github.com/abpoll/unsafe/issues?q=is%3Aissue%20state%3Aopen%20label%3A%22good%20first%20issue%22). If you find an issue to work on, you are welcome to open a PR with a fix.
+
+## Making Changes
+We welcome any contributions. You may find areas where code can be organized better or perform tasks more efficiently, documentation can be improved, and new features can be added. You may also find small issues with existing documentation, broken links, etc., We really appreciate your time in helping to make `UNSAFE` better. 
+
+Make a [fork](https://docs.github.com/en/github/getting-started-with-github/fork-a-repo#fork-an-example-repository) and from there create a working branch to get started on your modifications. When you are modifying any code, please rerun the partial and full data examples to ensure that existing functionality continues. In general, you do not have to redownload data or unzip raw data when doing these functionality tests if you have previously run these examples. 
+
+Commit your changes and submit a pull request once you verify that the partial and full data examples still work. 
