### 2025-09-13 abpoll/unsafe#13 @erexer
Looks great!
