### 2025-09-13 abpoll/unsafe#13 @erexer
shoot, looks like the links [here](https://github.com/abpoll/unsafe/tree/main/examples#getting-ready-to-run-the-examples) are missing a `?`, `https://htmlpreview.github.io/https://github.com/abpoll/unsafe/blob/main/examples/phil_frd_partial/notebooks/partial_data_example.html` needs to be `https://htmlpreview.github.io/?https://github.com/abpoll/unsafe/blob/main/examples/phil_frd_partial/notebooks/partial_data_example.html`
