### 2025-09-13 abpoll/unsafe#14 @erexer
Examples questions
Hi again, this is for the [JOSS review](https://github.com/openjournals/joss-reviews/issues/7527).

Hopefully this will be quick, thanks for working through this with me!!

I was able to run the partial data example and it looks like my results are slightly different than the [expected results](https://htmlpreview.github.io/?https://github.com/abpoll/unsafe/blob/main/examples/phil_frd_partial/notebooks/partial_data_example.html), does that track or should this be deterministic?

I'm also unsure about the full example, I've downloaded `dg.zip` and put it in the `unsafe/examples/philadelphia_frd/data/raw/external/haz/` dir. Does the ipynb download more data or is that all I need?

Thanks!
