### 2025-09-14 0afeb03
fix preview links and update context
