### 2025-09-14 abpoll/unsafe#13 @abpoll
Thanks so much for catching that! Fixed and added context about stochastic sampling and why users can't exactly verify these outputs: https://github.com/abpoll/unsafe/commit/0afeb0309bf500645ec8b7e704a3417956c8ebcf
