### 2025-09-15 abpoll/unsafe#14 @erexer
Ah ok, thanks for doing that!

It looks like my results for the full example are pretty far off from the [expected results](https://htmlpreview.github.io/?https://github.com/abpoll/unsafe/blob/main/examples/philadelphia_frd/notebooks/full_data_example.html), for the first graph the number of ensemble members are off by a factor of 10, do you know why that would be happening?
