+++ joss_submission/paper.md
@@ -11,7 +11,7 @@ authors:
     corresponding: true
     email: adam.b.pollack@dartmouth.edu
     orcid: 0000-0001-6642-0591
-    affiliation: [1, 2]
+    affiliation: "1, 2"
   - name: James Doss-Gollin
     orcid: 0000-0002-3428-2224
     affiliation: 3
