+++ joss_submission/paper.bib
@@ -15,6 +15,7 @@
   publisher = "Wiley Online Library",
   year      =  2019,
   url       = "https://agupubs.onlinelibrary.wiley.com/doi/abs/10.1029/2018EF001074",
+  doi       = "10.1029/2018ef001074",
   keywords  = "unsafe\_joss"
 }
 
