### 2025-10-15 #7527 @abpoll
Sorry 😅 Just updated. Citation on archive:

Pollack, A., Doss-Gollin, J., Srikrishnan, V., & Keller, K. (2025). UNSAFE: An UNcertain Structure And Fragility Ensemble framework for property-level flood risk estimation (joss2.1). Zenodo. https://doi.org/10.5281/zenodo.17316891
