### 2025-10-15 #7527 @atrisovic
Thank you! The version tag does not match GitHub. Can you make both eg v0.2.1?
