### 2025-10-15 a77f932
update readme with zenodo reference and get ready for v0.2.1 for joss
