### 2025-10-17 #7527 @atrisovic
Hi @abpoll

I read the paper closely and opened a PR here: https://github.com/abpoll/unsafe/pull/16/files

My other comments

1. In section Summary, can you add a subsection Example, and outline the example from the Jupyter notebook.
2. It's unclear what exact user inputs and outputs are, and those should be outlined in the example and before.
3. Add target audience and use case examples.
4. What is the spatial resolution and how it is set by the user.
5. What open data are used and how is uncertainty incorporated.
6. Jupyter notebooks should be referenced from Zenodo, not GH.
7. Add limitations as a stand alone subsection and outline limitations there.
