+++ joss_submission/paper.md
@@ -35,21 +35,138 @@ bibliography: paper.bib
 
 ---
 # Statement of Need
-Flooding is among the most frequent and damaging natural hazards in the United States [@Kousky2020]. Researchers and practitioners increasingly use flood-risk estimates to analyze dynamics and inform decisions [@Merz2010; @Trigg2016; @Bates2023; @Mulder2023]. There is increasing demand for flood-risk estimates at the scale of individual assets [@Condon2023]. One driver of this demand is that flood-risk estimates at coarser scales are susceptible to aggregation biases [@Pollack2022; @Condon2023]. Robust flood-risk estimates require explicit representation of uncertainties surrounding key inputs driving hazards, exposures, and vulnerabilities at relevant scales [@Bates2023; @Sieg2023; @Saint-Geours2015; @Tate2015; @Rozer2019; @Hosseini-Shakib2024].
+Flooding is among the most frequent and damaging hazards in the United States [@Kousky2020]. Researchers and practitioners increasingly use flood-risk estimates to analyze dynamics and inform decisions [@Merz2010; @Trigg2016; @Bates2023; @Mulder2023]. There is increasing demand for flood-risk estimates at the scale of individual assets [@Condon2023]. One driver of this demand is that flood-risk estimates at coarser scales are susceptible to aggregation biases [@Pollack2022; @Condon2023]. Robust flood-risk estimates require explicit representation of uncertainties surrounding key inputs driving hazards, exposures, and vulnerabilities at relevant scales [@Bates2023; @Sieg2023; @Saint-Geours2015; @Tate2015; @Rozer2019; @Hosseini-Shakib2024].
 
 Many property-level flood-risk estimation frameworks are silent on the effects of uncertainties surrounding key inputs [@Schneider2006; @Merz2010; @Saint-Geours2015; @Tate2015; @Sieg2023]. The predominant approach  to estimating economic flood damages in a U.S. setting relies on either the U.S. Army Corps of Engineers (USACE) or Federal Emergency Management Agency (FEMA) depth-damage functions (DDFs) [@Scawthorn2006; @Merz2010; @Tate2015]. In this DDF approach, flood-risk estimates depend on assumptions about several exposure and vulnerability characteristics. Risk estimates are sensitive to the spatial precision of linking a structure to a certain flood depth, the first-floor elevation of a structure, the structure's foundation type, the number of stories of the structure, the main function of the structure (i.e. residential or commercial), and the value of the structure [@Merz2010; @Tate2015; @Wing2020; @Pollack2022; @Xia2024]. Risk estimates are also sensitive to the expected damage for a given depth and the shape of the depth-damage relationship [@Merz2010; @Tate2015; @Rozer2019; @Wing2020]. All of these characteristics are subject to uncertainties which propagate to the resulting risk estimates [@Tate2015; @Saint-Geours2015; @Pollack2022; @Hosseini-Shakib2024].
 
-Here we implemented the Uncertain Structure and Fragility Ensemble (UNSAFE) framework to provide the U.S. flood-risk assessment community with a free and open-source tool for estimating property-level flood risks under uncertainty. UNSAFE represents exposure and vulnerability inputs under uncertainty using entirely free data. This extends the functionality of the most comparable publicly available tool, [“go-consequences”](https://github.com/USACE/go-consequences) [@USACE-go2024], developed by the USACE, for which limited documentation and usage examples are available. Examination of the repository suggests that go-consequences can produce stochastic representations of selected exposure and vulnerability characteristics in the DDF paradigm. However, this functionality is not available for key drivers like structure value or the functional form of the DDF for a given structure.  
[…]
