+++ joss_submission/paper.bib
@@ -744,3 +744,88 @@
   note         = "Accessed: 2024-3-5",
   keywords     = "unsafe\_joss"
 }
+
+@UNPUBLISHED{Pollack2025-nsi,
+  title    = "Unrefined national building inventories can mislead risk
+              assessments and decisions",
+  author   = "Pollack, Adam and Benedict, James and Deb, Mithun and Doss-Gollin,
+              James and Judi, David and Lehman, William and Lutz, Nicholas and
+              Reesman, Cade and Sarazen, Elaine and Son, Youngjun and
+              Srikrishnan, Vivek and Sun, Ning and Keller, Klaus",
+  journal  = "SSRN Electronic Journal",
+  abstract = "Flood-risk assessments inform consequential public decisions.
+              These assessments increasingly use large-scale building
+              inventories that offer granular details bu",
+  month    =  oct,
+  year     =  2025,
+  url      = "https://papers.ssrn.com/abstract=5575271",
+  keywords = "flood risk, national structure inventory, resource allocation"
+}
+
+@ARTICLE{Pollack2025-j40,
+  title     = "Funding rules that promote equity in climate adaptation outcomes",
+  author    = "Pollack, Adam B and Santamaria-Aguilar, Sara and Maduwantha,
+               Pravin and Helgeson, Casey and Wahl, Thomas and Keller, Klaus",
+  journal   = "Proceedings of the National Academy of Sciences",
+  publisher = "National Academy of Sciences",
+  volume    =  122,
+  number    =  2,
+  pages     = "e2418711121",
+  abstract  = "Many climate policies adopt improving equity as a key objective.
+               A key challenge is that policies often conceive of equity in
+               terms of individuals but introduce strategies that focus on
+               spatially c...",
+  month     =  jan,
+  year      =  2025,
+  url       = "https://www.pnas.org/doi/abs/10.1073/pnas.2418711121",
+  doi       = "10.1073/pnas.2418711121",
+  language  = "en"
+}
+
+@ARTICLE{Bhaduri2025,
+  title     = "Uncertainty in household behavior drives large variation in the
+               size of the levee effect",
+  author    = "Bhaduri, Parin and Pollack, Adam B and Yoon, James and Roy
+               Chowdhury, Pranab K and Wan, Heng and Judi, David and Daniel,
+               Brent and Srikrishnan, Vivek",
+  journal   = "J. Flood Risk Manag.",
+  publisher = "Wiley",
+  volume    =  18,
+  number    =  4,
+  pages     = "e70131",
+  abstract  = "ABSTRACTCoastal cities face increasing flood hazards due to
+               climate change. Physical infrastructures, such as levees, are
+               commonly used to reduce flood hazards. To effectively manage
+               flood risks, it is important to understand the degree to which
+               physical infrastructures change both hazard and exposure. For
+               example, many studies suggest that levee construction causes an
+               overall increase in risk because levees promote exposure growth
+               to a greater degree than they reduce flood hazards. Although this
+               so‐called “levee effect” is widely studied, there are knowledge
+               gaps surrounding how uncertainties related to levee construction
+               and flood risk translate into the occurrence and strength of the
+               levee effect in coastal communities. Here, we use agent‐based
+               modeling to simulate the influence of flood risk information
+               pathways on the dynamics around the levee effect, first under
+               idealized conditions and then within a real‐world coastal case
+               study. We finally conduct a global sensitivity analysis to
+               identify which model factors contribute to the strength of the
+               levee effect. We find that, under idealized conditions, the
+               strength of the levee effect is highly sensitive to economic
+               (e.g., population growth) and engineering (e.g., levee failure)
+               factors. However, under more complex coastal conditions, factors
[…]
