+++ joss_submission/paper.bib
@@ -748,7 +748,7 @@
 @UNPUBLISHED{Pollack2025-nsi,
   title    = "Unrefined national building inventories can mislead risk
               assessments and decisions",
-  author   = "Pollack, Adam and Benedict, James and Deb, Mithun and Doss-Gollin,
+  author   = "Pollack, Adam B and Benedict, James and Deb, Mithun and Doss-Gollin,
               James and Judi, David and Lehman, William and Lutz, Nicholas and
               Reesman, Cade and Sarazen, Elaine and Son, Youngjun and
               Srikrishnan, Vivek and Sun, Ning and Keller, Klaus",
