+++ joss_submission/paper.md
@@ -123,7 +123,7 @@ def get_loss_ensemble(
     """
 ```
 
-This function returns a DataFrame containing an ensemble of damage estimates and realizations of uncertain characteristics n_sow realizations for a sample size of *n_sow*. After running this function, a user can conduct a wide range of analyses (e.g., as done in [@Bhaduri2025 or @Pollack2025-nsi] though it is also possible to use `UNSAFE` to obtain more robust estimates of the mean damage estimate for a structure across the sampled realizations (e.g., [@Pollack2025-j40])
+This function returns a DataFrame containing an ensemble of damage estimates and realizations of uncertain characteristics n_sow realizations for a sample size of *n_sow*. After running this function, a user can conduct a wide range of analyses (e.g., as done in [@Bhaduri2025 or @Pollack2025-nsi]) though it is also possible to use `UNSAFE` to obtain more robust estimates of the mean damage estimate for a structure across the sampled realizations (e.g., [@Pollack2025-j40])
 
 The function includes default values for the options in the `config` dict argument. The default behavior for the function infers the uncertain distribution for parameters such as a structure's number of stories or foundation type based on overall census tract proportions. Users may modify these priors with the keys `stories_param` or `found_param`. Users may modify other priors as well. For example, users can update uncertainty around structure value assessment by modifying the `coef_var` key, or first-floor elevation with structure-level updates or a dict that maps foundation types to a distribution of possible foundation heights. 
 
