+++ joss_submission/paper.md
@@ -44,7 +44,7 @@ Here we implemented the Uncertain Structure and Fragility Ensemble (UNSAFE) fram
 Several existing tools help contextualize the methodological gap that UNSAFE addresses. The most prominent is [Hazus](https://www.fema.gov/flood-maps/products-tools/hazus) [@Scawthorn2006; @Schneider2006; @Tate2015], a freely available GIS-based desktop application for Windows that supports deterministic flood-damage assessments but cannot be readily modified to incorporate uncertainty in exposure and vulnerability. FEMA also developed the [“Flood Assessment Structure Tool”](https://github.com/nhrap-hazus/FAST?tab=readme-ov-file) [@FEMA2021] to facilitate more efficient deterministic Hazus analyses in Python, although this tool appears to be deprecated. The USACE maintains two published tools for deterministic analyses,([HEC-FIA](https://www.hec.usace.army.mil/confluence/fiadocs/fiaum/latest) [@USACE2021] and [HEC-FDA](https://www.hec.usace.army.mil/software/hec-fda/documentation/CPD-72_V1.4.1.pdf) [@USACE2024]). Lastly, there is the [Delft-FIAT (Fast Impact Assessment Tool)](https://deltares.github.io/Delft-FIAT/stable/) [@Deltares2024], developed and maintained by Deltares. It currently does not accommodate flood risk estimation with uncertainty in exposure and vulnerability inputs, but is open-source and well-documented. Advanced users could, in principle, extend its functionality to represent such uncertainty. `UNSAFE` provides a direct and streamlined framework for doing so.
 
 # Summary
-`UNSAFE` adopts and expands on a property-level risk assessment framework common in academic research and practice (e.g., [Federal Emergency Management Agency (FEMA) loss avoidance studies](https://www.fema.gov/grants/mitigation/loss-avoidance-studies) [@FEMA2024], [United States Army Corps of Engineers (USACE) feasibility studies](https://www.nad.usace.army.mil/Portals/40/docs/NACCS/10A_PhysicalDepthDmgFxSummary_26Jan2015.pdf) [@USACE2015]). UNSAFE allows users to add parametric uncertainty to building inventories, such as the widely used [National Structure Inventory (NSI) dataset](https://www.hec.usace.army.mil/confluence/nsi/technicalreferences/2019/technical-documentation) [@USACE-nsi2024] (i.e. uncertainty in exposure), and facilitates the use of multiple, potentially conflicting, expert-based DDFs (i.e. deep uncertainty in vulnerability). As of now, `UNSAFE` includes functionality to conduct risk assessments on residential properties of at most 3 stories given limitations in representing uncertainty in depth-damage relationships for other structures. This subset of structures represents a large proportion of structures exposed to flooding, and is a common inventory focus in research [@Pollack2025-nsi]. 
[…]
