### 2025-10-23 c0422af
polishing text and updating limitations section again
