### 2025-11-12 #7527 @atrisovic
Hi @abpoll! I simplified this version by removing some of the extra technical details :)
Could you please take a look at my PR? https://github.com/abpoll/unsafe/pull/18
Sorry for any earlier confusion!
