### 2025-11-13 #7527 @editorialbot
:wave: @openjournals/dsais-eics, this paper is ready to be accepted and published.
