### 2025-11-13 4261e28
add curly braces to enforce capitalization
