+++ joss_submission/paper.bib
@@ -5,7 +5,7 @@
 % to a LaTeX equivalent.
 @ARTICLE{Rozer2019,
   title     = "Probabilistic models significantly reduce uncertainty in
-               Hurricane Harvey pluvial flood loss estimates",
+               {H}urricane {H}arvey pluvial flood loss estimates",
   author    = "R{\"o}zer, V and Kreibich, H and Schr{\"o}ter, K and M{\"u}ller,
                M and {others}",
   abstract  = "Pluvial flood risk is mostly excluded in urban flood risk
@@ -165,7 +165,7 @@
 @ARTICLE{Saint-Geours2015,
   title     = "Ranking sources of uncertainty in flood damage modelling: a case
                study on the cost‐benefit analysis of a flood mitigation project
-               in the Orb Delta, France",
+               in the {O}rb {D}elta, {F}rance",
   author    = "Saint-Geours, N and Grelot, F and Bailly, J-S and Lavergne, C",
   abstract  = "AbstractCost‐benefit analyses (CBA) of flood management plans
                usually require estimating expected annual flood damages on a
@@ -473,7 +473,7 @@
 
 @ARTICLE{Pollack2022,
   title     = "Aggregation bias and its drivers in large scale flood loss
-               estimation: A Massachusetts case study",
+               estimation: A {M}assachusetts case study",
   author    = "Pollack, Adam B and Sue Wing, Ian and Nolte, Christoph",
   abstract  = "Abstract Large-scale estimations of flood losses are often based
                on spatially aggregated inputs. This makes risk assessments
@@ -688,7 +688,7 @@
   author      = "{United States Army Corps of Engineers}",
   abstract    = "A lightweight consequences computational engine written in Go
                  - USACE/go-consequences",
-  institution = "Github",
+  institution = "Git{H}ub",
   year        =  2024,
   url         = "https://github.com/USACE/go-consequences",
   keywords    = "unsafe\_joss",
@@ -724,7 +724,7 @@
   author      = "{Federal Emergency Management Agency}",
   abstract    = "Hazus - Flood Assessment Structure Tool . Contribute to
                  nhrap-hazus/FAST development by creating an account on GitHub.",
-  institution = "Github",
+  institution = "Git{H}ub",
   month       =  mar,
   year        =  2021,
   url         = "https://github.com/nhrap-hazus/FAST",
@@ -733,7 +733,7 @@
 }
 
 @MISC{USACE2015,
-  title        = "North Atlantic Coast Comprehensive Study: Resilient
+  title        = "North {A}tlantic {C}oast Comprehensive Study: Resilient
                   Adaptation to Increasing Risk, Physical Depth Damage Function
                   Summary Report",
   author       = "{United States Army Corps of Engineers}",
